Extract the delete confirmation modal into `plottr_components`
