+++ lib/plottr_components/package.json
@@ -17,7 +17,7 @@
     "web-vitals": "^1.0.1"
   },
   "scripts": {
-    "start": "npm run clean && cross-env NODE_ENV=production babel src/components --watch --out-dir dist --copy-files --ignore __tests__,spec.js,test.js,stories.js,__snapshots__",
+    "start": "cross-env NODE_ENV=production babel src/components --watch --out-dir dist --copy-files --ignore __tests__,spec.js,test.js,stories.js,__snapshots__",
     "eject": "react-scripts eject",
     "clean": "rimraf dist",
     "compile": "npm run clean && cross-env NODE_ENV=production babel src/components --out-dir dist --copy-files --ignore __tests__,spec.js,test.js,stories.js,__snapshots__",
@@ -57,6 +57,8 @@
   "peerDependencies": {
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "react-scripts": "4.0.3"
+    "react-scripts": "4.0.3",
+    "plottr_locales": "^0.0.1",
+    "pltr": "^2.2.11"
   }
 }
