+++ lib/plottr_components/src/components/DeleteConfirmModal.js
