+++ lib/plottr_components/src/components/index.js
@@ -1,5 +1,5 @@
-import Button from './Button'
+import DeleteConfirmModal from './DeleteConfirmModal'
 
 export {
-  Button
+  DeleteConfirmModal
 }
