+++ src/app/components/project/Book.js
@@ -8,7 +8,7 @@ import Image from '../images/Image'
 import BookDialog from './BookDialog'
 import { Glyphicon, ButtonGroup, Button } from 'react-bootstrap'
 import cx from 'classnames'
-import DeleteConfirmModal from '../dialogs/DeleteConfirmModal'
+import { DeleteConfirmModal } from 'plottr_components'
 import { actions, selectors } from 'pltr/v2'
 
 const BookActions = actions.book
