+++ src/app/components/tag/tagView.js
@@ -12,7 +12,7 @@ import {
   Glyphicon,
 } from 'react-bootstrap'
 import ColorPicker from '../colorpicker'
-import DeleteConfirmModal from '../dialogs/DeleteConfirmModal'
+import { DeleteConfirmModal } from 'plottr_components'
 import { t as i18n } from 'plottr_locales'
 import cx from 'classnames'
 import { actions } from 'pltr/v2'
