+++ src/app/components/timeline/BeatTitleCell.js
@@ -13,7 +13,7 @@ import {
 } from 'react-bootstrap'
 import { Cell } from 'react-sticky-table'
 import cx from 'classnames'
-import DeleteConfirmModal from '../dialogs/DeleteConfirmModal'
+import { DeleteConfirmModal } from 'plottr_components'
 import { actions, helpers, selectors } from 'pltr/v2'
 import InputModal from '../dialogs/InputModal'
