+++ src/app/components/timeline/CardDialog.js
@@ -20,7 +20,7 @@ import cx from 'classnames'
 import tinycolor from 'tinycolor2'
 import RichText from '../rce/RichText'
 import SelectList from 'components/selectList'
-import DeleteConfirmModal from '../dialogs/DeleteConfirmModal'
+import { DeleteConfirmModal } from 'plottr_components'
 import EditAttribute from '../EditAttribute'
 import { helpers, actions, selectors } from 'pltr/v2'
 import ColorPickerColor from '../ColorPickerColor'
