+++ src/app/components/timeline/ClearNavItem.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { t as i18n } from 'plottr_locales'
 import { NavItem, Button, Glyphicon } from 'react-bootstrap'
-import DeleteConfirmModal from '../dialogs/DeleteConfirmModal'
+import { DeleteConfirmModal } from 'plottr_components'
 import { selectors, actions } from 'pltr/v2'
 
 const { currentTimelineSelector } = selectors
